requirement.txt
mysql-connector-python
python-dotenv
langchain
langchain-community
langchainhub
openai
ctransformers
requests
streamlit
llama-index